Correct Camera Rotation
The issue wasn't that the vectors needed to be normalized, it
was just that the dir/cam values were being changed *and* those
new values were being used in the calculations.

This commit fixes that to copy the dir/cam vectors so we have an
immutable copy, then change the assignments to just point straight
to the camera's vector members.