Add Initial Weapon Inventory
This adds a basic inventory and method for the planned weapon types.

Currently all types use a temporary texture for their HUD display, and
shooting is not yet enabled. Weapon sway and bobbing are added and
affect the weapon HUD texture.

This also adds keyboard keys for jumping and sprinting, which were
previously only mapped for game controllers.