Update Jump To Be Z-Velocity Tolerance Based
Rather than requiring the player to be standing on a surface, this
commit updates jump logic so that there is a small amount of time
that the player is allowed to jump after stepping off of a surface.
This slightly emulates having one foot still on the surface and
pushing up and off with it.