Add Collision Detection Tables
This adds in sets for tracking entity collisions between frames. When
two entities collide their IDs are recorded to the table, but only once
per frame. Previous frame collisions are stored so that collision logic
can check if the collision is "initial" or not (used for things like
damage).

This also updates the naming of some of the Fang_Body functions to keep
them more consistent.