Convert Fang_Body To Interpolated Velocity
This removes several redundant fields on Fang_Body and consolidates
the approach to movement. Bodies will still maintain a direction but
are now responsible for calculating the intended velocity themselves.
This means that bodies can now move independently of their facing
direction.

The 'size' member was also replaced with width and height so that
bodies can maintain both the radii of their collision cylinder as
well as the height.