Add Map Collision Callbacks For Entities
This moves the map collision detection into its own function so
that it can run independent of the entity movement. Now the engine
first moves entities and then checks their collisions with the map
afterwards.

If a collision occured the entity's map collision callback is called,
however the position resolution is maintained by the physics system.
Entities cannot currently modify the behavior of stopping at walls
on collision.