Add Fang_PlayerFireWeapon()
This moves the logic for the player to fire a weapon into its own function,
so that the input handling code doesn't have to be concerned with weapon
logic.

This also sets placeholder damage values for the current weapon types in
the game. The damage is used to set the projectile damage but is also used
for the health of projectiles as well.