Move Camera Pitch To Direction Vector
The pitch was previously being stored in the 'cam' (camera plane)
vector, which was not really representative of the camera's direction. It
also meant that the Z-direction was not available to the physics system.

This also corrects Fang_BodySetTargetVelocity(). This function shouldn't need
to take an 'up' value, and should take the body Z-direction into account
when normalizing the direction vectors. With this change, projectiles can
travel upwards as well as forward.

For bodies that utilize the jumping ability, they should manually set their
target Z-velocity after Fang_SetTargetVelocity().