Transition To Chunks For Map Storage
This moves the map data away from the simple tile array to a
chunk storage system. Chunks are 16x16 units in size, with 2^6 total
chunks in the world. A chunked world map allows for a few immediate and
future features:

- Spatial hashing for entities and tiles
- Reducing operations needed during tile/entity collision checks
- Limiting entity updates based on chunk positions
- Allowing different floor textures per chunk

Additionally this commit moves all of the collision related functions out
of the Fang_Entity.c file. Ideally all files included prior to the main
logic (currently in Fang.c) should be as small in scope as possible. They
should declare types they use and the logical functions associated with
those types, but should not be responsible for core "game logic". This will
hold true especially when other modes, such as the edit mode, are added, as
other modes may need to alter the behavior of the core game pieces (collision,
physics, etc.).