Convert World To Chunk Based Storage
This moves the map data away from the simple tile array to a
chunk storage system. Chunks are 16x16 tiles in size, with 2^6 total
chunks in the world. A chunked world map allows for a few immediate and
future features:

- Spatial hashing for entities and tiles
- Reducing operations needed during tile/entity collision checks
- Limiting entity updates based on chunk positions
- Allowing different floor textures per chunk

This commit also moves all of the collision related functions out
of the Fang_Entity.c file. Ideally all files included prior to the main
logic (currently in Fang.c) should be as small in scope as possible. 

Additionally a stack-size bug regarding the initialization of the world 
map has been fixed. Compiler diagnostics are now enabled to catch
such problems at compile time.