Add Basic Audio Playback
This adds the ability to play audio (either sound effects or music tracks) in
the game via an audio callback method.

Audio files are loaded and then played in the game via Fang_Sound. Sounds can
be "global" or "positional". Positional sounds are panned according to the
"listener" position and direction, giving the effect of directional audio.

Audio files played as music are queued such that they fade in and out when the
queued music type changes.